// Fontend/TNT.N8.Web/TNT.N8.Web/ClientApp/src/app/product/models/product.model.ts
export class EmployeeEntityModel {
  employeeId: string;
  mission: string;
  employeeName: string;
  employeeCode: string;
  OrganizationId: string;
  PositionId: string;
  EmployeeCodeName: string;
  roleId : string | null;
  isDisable : boolean | null;
  value : string | null;
  chucVuId: number;
}
// ...
export class CustomerOrderExtension {
  id: string | null;
  dataType: number | null;
  attributeConfigurationId: string;
  dataTypeName: string;
  categoryId: string | null;
  categoryName: string;
  createdById: string | null;
  createdDate: string | null;
  updatedById: string | null;
  updatedDate: string | null;
  tenantId: string | null;
  objectType: number | null;
  objectId: string | null;
  value: any;
  orderDetailId: any;
  servicePacketMappingOptionsId: any; 
  attrName: any; 
}