// Fontend/TNT.N8.Web/TNT.N8.Web/ClientApp/src/app/order/models/customer-order.model.ts
export class CustomerOrderTaskEntityModel {
  id: string;
  orderDetailId: string | null;
  optionId: string;
  optionName: string;
  path: string;
  vendorId: string | null;
  vendorName: string;
  empId: string | null;
  listEmpId: Array<string> | null;
  listEmpName: string | null;
  empName: string;
  mission: string;
  note: string;
  servicePacketId: string;
  isExtend: boolean;
  createdById: string;
  createdDate: string;
  updatedById: string;
  updatedDate: string;
  tenantId: string;
  isPersonInCharge: boolean;
  orderActionId: string;
  nguoiThucHienType: number;
  hanPheDuyet: Date;
  level: number;
  statusId: number

  vendorOrderId: string;
  vendorOrderCode: string;
  enableBaoCaoDoanhThu: boolean;

  showBtn: boolean;
  status: string;
}